src/delivery_app/requirements.txt
streamlit==1.11.1
mysql-connector-python
altair==4

requests>2.23.0,<=2.31.0

